Refactor isDisplayed atom to remove Closure library code
The Closure library adds a lot of complexity to the compiled and minified atoms,
including for isDisplayed. Additionally, the web platform has evolved in the
time since the atoms were originally written, with additional methods added that
help simplify the atom. This commit removes (nearly) all of the Closure code
from the specific atom used by the isDisplayed method in the various language
bindings.